+++ Blish HUD/GameServices/Overlay/UI/Views/OverlaySettingsView.cs
@@ -29,6 +29,7 @@ namespace Blish_HUD.Overlay.UI.Views {
         private void BuildOverlaySettings(Panel rootPanel) {
             GetStandardPanel(rootPanel, Strings.GameServices.OverlayService.OverlaySettingsSection).Show(new SettingsView(GameService.Overlay.OverlaySettings));
             GetStandardPanel(rootPanel, Strings.GameServices.GraphicsService.GraphicsSettingsSection).Show(new SettingsView(GameService.Graphics.GraphicsSettings));
+            GetStandardPanel(rootPanel, Strings.GameServices.DebugService.DebugSettingsSection).Show(new SettingsView(GameService.Debug.DebugSettings));
         }
 
     }
